test(testing): sync test support вынесен в отдельные файлы
CrossDeviceReadingScenarioTests и общий RSSReaderTestSupport выделены из RSSReaderTests.swift, чтобы sync-сценарии и shared test helpers не жили в одном монолитном test file
